requirements.txt
flask==2.3.3
sqlalchemy==2.0.19
python-dotenv==1.0.0
requests==2.31.0
flask-sqlalchemy==3.0.5
werkzeug==2.3.7
gunicorn==21.2.0
alembic==1.12.0 